[Author: thomasvl]
Add another window to the UILocalizer unittest that puts a bunch of
  items into boxes, splitters, and scroll views to make sure they are
  working.
Support some NSBox attributes in AppKit+UnitTesting.
Update the size to fit code to only special case regular push buttons and
  not other types of buttons.
Update the expectations due to this change.
Add a new unittest to test all button styles and sizes.
Update how we decided to do vertical or horizontal layout on a width box
  because left aligning things is visual and doesn't actually align their
  frames, so instead we do a rect intersection instead.
Added in the CompilerSDK on Mac OS builds to the paths searches.
Update the comments on how paths are searched to be correct.

R=dmaclach,stuartmorgan
DELTA=2709  (2643 added, 8 deleted, 58 changed)

git-svn-id: http://google-toolbox-for-mac.googlecode.com/svn/trunk@197 7dc7ac4e-7543-0410-b95c-c1676fc8e2a3