[Author: thomasvl]
Don't trim the whitespace off the string when doing test sizing...
TextField does some odd stuff with whitespace when it does wrap things, if you don't include the white space, then when you use the real string (to wrap), it moves the last word down a line because of lack of space for the whitespace (go figure).
Oh, and fix the leak.

R=dmaclach
DELTA=6  (0 added, 1 deleted, 5 changed)

git-svn-id: http://google-toolbox-for-mac.googlecode.com/svn/trunk@304 7dc7ac4e-7543-0410-b95c-c1676fc8e2a3