[Author: avi]
Allow adding a new sheet in a sheet's closing callback in GTMWindowSheetController. GTMWindowSheetController does not completely remove a sheet until after the callback has been invoked. This causes an assert to trigger if another sheet is launched in the callback, which is inconsistent with normal Cocoa sheets. Switch the order of the two operations and add a unit test. This affects http://crbug.com/56948 . Patch by davidben@chromium.org .

R=dmaclach
DELTA=112  (106 added, 5 deleted, 1 changed)

git-svn-id: http://google-toolbox-for-mac.googlecode.com/svn/trunk@410 7dc7ac4e-7543-0410-b95c-c1676fc8e2a3