[Author: thomasvl]
Take 2: Overhaul how iOS unittests are run.

Newer versions of Xcode have changes in the Simulator and supporting
frameworks so the way GTM had been running unittests no longer works.
The Simulator frameworks bail on launch with a zero exit code, so unless
you look at the raw output, you don't even notice the tests no longer
run.

This new way comes from the work done in Chromium to actually launch the
unittests fully under the simulator to get them a much more realistic
environment. It also takes what was learned there about getting the output
of the binary and how to deal with all the different ways a test and the
simulator can fail to start along with all the different ways the failures/
crashing of the app can come back.

There are some new env variables that let you force a version of the
simulator (assuming it is installed) along with one to say if you want to
run under iPhone or iPad.

- Land a copy of the Chromium iossim util.
- Land RuniOSUnitTestsUnderSimulator that uses iossim, projects have to be
  moved over to use it.

R=dmaclach
DELTA=227  (227 added, 0 deleted, 0 changed)

git-svn-id: http://google-toolbox-for-mac.googlecode.com/svn/trunk@567 7dc7ac4e-7543-0410-b95c-c1676fc8e2a3